Fix a subtle NaN issue.
Sigmoidal remapping can cause floating point infinities to turn up in the intermediate
representation. Most of the algorithm has no problem with that, however that specific
line would convert them to NaNs, because ∞-∞ = NaN in IEEE-754.
This trivial reformulation (equivalency in ℝ) avoids the problem, at least as long as
no two opposite infinities appear in the same expression.